+++ tests/compile-fail/UncarryableEnumOptionalDefault.kt
@@ -1,0 +1,3 @@
+import UncarryableEnumOptionalDefault.Cases
+
+fun uncarryableEnumOptionalDefault() = Cases.Read()